backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
yfinance==0.2.61
pandas==2.2.3
pydantic==1.10.26
python-dotenv==1.2.1
PyYAML==6.0.3
